fix: useHydrateCache always restores cache, logger is direct import
