+++ packages/next/src/HeadlessProvider.tsx
@@ -1,4 +1,4 @@
-import isNull from 'lodash/isNull';
+import isNil from 'lodash/isNil';
 import React from 'react';
 import { client } from './client';
 import { CLIENT_CACHE_PROP, PageProps } from './getProps';
@@ -15,9 +15,12 @@ export function HeadlessProvider<Props = Record<string, unknown>>({
   const { useHydrateCache } = client();
   const cacheSnapshot = pageProps[CLIENT_CACHE_PROP];
 
+  if (typeof window !== 'undefined') {
+    console.log(JSON.parse(cacheSnapshot ?? '{}'));
+  }
+
   useHydrateCache({
-    cacheSnapshot: isNull(cacheSnapshot) ? undefined : cacheSnapshot,
-    shouldRefetch: false,
+    cacheSnapshot: isNil(cacheSnapshot) ? undefined : cacheSnapshot,
   });
 
   return (
